# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API for cosbi financial site",
    "description": "API document for cosbi financial site made by express",
    "version": "1.0.0"
  },
  "host": "140.116.214.154:3000",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "swagger_token"
    }
  },
  "paths": {
...
    "/swagger/token": {
      "get": {
        "tags": [
          "Token"
        ],
        "description": "Get swagger token.",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "description": "Swagger username.",
            "required": true,
            "type": "string",
            "schema": {
              "type": "string",
              "example": "example_swagger_username"
            }
          },
          {
            "name": "password",
            "in": "query",
            "description": "Swagger password.",
            "required": true,
            "type": "string",
            "schema": {
              "type": "string",
              "example": "example_swagger_password"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "User login.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Username and password.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userName": {
                  "type": "string",
                  "example": "example_username"
                },
                "password": {
                  "type": "string",
                  "example": "example_passord"
                }
              },
              "required": [
                "userName",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/user/register": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Register user.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Register user.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "test"
                },
                "userName": {
                  "type": "string",
                  "example": "test123"
                },
                "email": {
                  "type": "string",
                  "example": "test123@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "Test123456"
                }
              },
              "required": [
                "name",
                "userName",
                "email",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/user/logout": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "User logout.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/data/isAuth": {
      "get": {
        "tags": [
          "Authenticate check"
        ],
        "description": "Check if user have auth.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/data/download/single_financialData": {
      "get": {
        "tags": [
          "File download"
        ],
        "description": "Download ticker research pdf.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "filename",
            "in": "query",
            "description": "Filename about ticker research pdf.",
            "required": true,
            "type": "string",
            "schema": {
              "type": "string",
              "example": "6277_宏正_20230419_宏遠_區間操作_NULL.pdf"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/data/download/single_post_board_memo": {
      "get": {
        "tags": [
          "File download"
        ],
        "description": "Download post board memo data.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "filename",
            "in": "query",
            "description": "Filename about post board memo.",
            "required": true,
            "type": "string",
            "schema": ""
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/data/download/single_line_memo": {
      "get": {
        "tags": [
          "File download"
        ],
        "description": "Download line memo data.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "filename",
            "in": "query",
            "description": "Filename about line memo.",
            "required": true,
            "type": "string",
            "schema": {
              "type": "string",
              "example": "8210_勤誠_2023-02-07_16_59_11.txt"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/data/download/single_twse_chPDF": {
      "get": {
        "tags": [
          "File download"
        ],
        "description": "Download twse ch pdf.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "table_name",
            "in": "query",
            "description": "Filename about twse ch pdf",
            "required": true,
            "type": "string",
            "schema": {
              "type": "string",
              "example": "247220230418M001.pdf"
            }
          },
          {
            "name": "filename",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/data/download/single_twse_enPDF": {
      "get": {
        "tags": [
          "File download"
        ],
        "description": "Download twse en pdf.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "table_name",
            "in": "query",
            "description": "Filename about twse en pdf",
            "required": true,
            "type": "string",
            "schema": {
              "type": "string",
              "example": "247220230418E001.pdf"
            }
          },
          {
            "name": "filename",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/data/download/single_financialDataOther": {
      "get": {
        "tags": [
          "File download"
        ],
        "description": "Download other research pdf.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "filename",
            "in": "query",
            "description": "Filename about other research pdf.",
            "required": true,
            "type": "string",
            "schema": {
              "type": "string",
              "example": "SinoPac股市重大事件行事曆-04182023.PDF"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/data/download/single_financialDataIndustry": {
      "get": {
        "tags": [
          "File download"
        ],
        "description": "Download industry research pdf.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "filename",
            "in": "query",
            "description": "Filename about industry research pdf.",
            "required": true,
            "type": "string",
            "schema": {
              "type": "string",
              "example": "投資快訊03092023-半導體產業-存貨沒意外地續揚.pdf"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/data/download/table_status": {
      "get": {
        "tags": [
          "File download"
        ],
        "description": "Download sql table as csv file.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "table_name",
            "in": "query",
            "description": "Table name and add .csv.",
            "required": true,
            "type": "string",
            "schema": {
              "type": "string",
              "example": "news.csv"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/data/upload/post_board_upload": {
      "post": {
        "tags": [
          "File upload"
        ],
        "description": "Upload post board memo.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "Post board memo parameter\n   ### Middleware-This API uses multer middleware to handle upload file.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "date": {
                  "type": "string",
                  "example": "2023-04-19"
                },
                "stock_num_name": {
                  "type": "string",
                  "example": "2330 台積電"
                },
                "recommend": {
                  "type": "string",
                  "example": "買進"
                },
                "price": {
                  "type": "string",
                  "example": "500"
                },
                "reason": {
                  "type": "string",
                  "example": "原因..."
                },
                "filename": {
                  "type": "string",
                  "example": "123.txt"
                }
              },
              "required": [
                "date",
                "stock_num_name",
                "recommend",
                "price",
                "reason",
                "filename"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/data/upload/line_memo_upload": {
      "post": {
        "tags": [
          "File upload"
        ],
        "description": "Upload line memo.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "Line memo research parameter.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "date": {
                  "type": "string",
                  "example": "2023-04-19"
                },
                "stock_num_name": {
                  "type": "string",
                  "example": "2330 台積電"
                },
                "content": {
                  "type": "string",
                  "example": "line memo 內容"
                }
              },
              "required": [
                "date",
                "stock_num_name",
                "content"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/data/upload/self_upload": {
      "post": {
        "tags": [
          "File upload"
        ],
        "description": "Upload ticker research pdf.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "Ticker research parameter.\n   ### Middleware-This API uses multer middleware to handle upload file.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "ticker": {
                  "type": "string",
                  "example": "2330 台積電"
                },
                "date": {
                  "type": "string",
                  "example": "2023-04-19"
                },
                "provider": {
                  "type": "string",
                  "example": "統一投顧"
                },
                "evaluate": {
                  "type": "string",
                  "example": "買進"
                }
              },
              "required": [
                "ticker",
                "date",
                "provider",
                "evaluate"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/data/upload/other": {
      "post": {
        "tags": [
          "File upload"
        ],
        "description": "Upload other research pdf.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "Other research parameter\n   ### Middleware-This API uses multer middleware to handle upload file.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "date": {
                  "type": "string",
                  "example": "2023-04-19"
                },
                "title": {
                  "type": "string",
                  "example": "台積電..."
                },
                "investmentCompany": {
                  "type": "string",
                  "example": "統一投顧"
                },
                "filname": {
                  "type": "string",
                  "example": "台積電???.pdf"
                }
              },
              "required": [
                "date",
                "title",
                "investmentCompany",
                "filname"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/data/upload/industry": {
      "post": {
        "tags": [
          "File upload"
        ],
        "description": "Upload industry research pdf.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "Industry research parameter\n   ### Middleware-This API uses multer middleware to handle upload file.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "date": {
                  "type": "string",
                  "example": "2023-04-19"
                },
                "title": {
                  "type": "string",
                  "example": "台積電..."
                },
                "category": {
                  "type": "string",
                  "example": "半導體"
                },
                "investmentCompany": {
                  "type": "string",
                  "example": "統一投顧"
                },
                "filname": {
                  "type": "string",
                  "example": "台積電???.pdf"
                }
              },
              "required": [
                "date",
                "title",
                "category",
                "investmentCompany",
                "filname"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/data/newestNews20": {
      "get": {
        "tags": [
          "Get data"
        ],
        "description": "Get newest 20 news.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/data/newestResearch20": {
      "get": {
        "tags": [
          "Get data"
        ],
        "description": "Get newest 20 ticker research.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/data/table_status": {
      "get": {
        "tags": [
          "Get data"
        ],
        "description": "Get table status.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/data/post_board_state": {
      "get": {
        "tags": [
          "Get data"
        ],
        "description": "Get post board memo table status.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/data/lineMemo_state": {
      "get": {
        "tags": [
          "Get data"
        ],
        "description": "Get line memo table status.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/data/superUser": {
      "get": {
        "tags": [
          "Get data"
        ],
        "description": "Check if user is super user.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/data/userList": {
      "get": {
        "tags": [
          "Get data"
        ],
        "description": "Get user list.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/data/username": {
      "get": {
        "tags": [
          "Get data"
        ],
        "description": "Get username.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/data/calender": {
      "post": {
        "tags": [
          "Get data"
        ],
        "description": "Get twse data and transorm to calender type.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "Twse year and month.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "year": {
                  "type": "string",
                  "example": "2023"
                },
                "month": {
                  "type": "string",
                  "example": "04"
                }
              },
              "required": [
                "year",
                "month"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/data/calenderData": {
      "post": {
        "tags": [
          "Get data"
        ],
        "description": "Get twse data.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "Twse year and month.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "year": {
                  "type": "string",
                  "example": "2023"
                },
                "month": {
                  "type": "string",
                  "example": "04"
                }
              },
              "required": [
                "year",
                "month"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/data/ticker_list": {
      "get": {
        "tags": [
          "Get data"
        ],
        "description": "4 number ticker list.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
...
    "/data/inflation": {
      "get": {
        "tags": [
          "Python tool"
        ],
        "description": "Inflation.",
        "parameters": [
          {
            "name": "swagger_token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
...
  },
...
}